+++ apps/sitecontent/models.py
@@ -1,0 +1,6 @@
+from wagtail.models import Page
+
+
+class HomePage(Page):
+    parent_page_types = ["wagtailcore.Page"]
+    subpage_types: list[str] = []