+++ framehold/settings/test.py
@@ -1,0 +1,18 @@
+from .base import *  # noqa: F403
+
+DEBUG = False
+ALLOWED_HOSTS = ["testserver", "localhost", "127.0.0.1"]
+CSRF_TRUSTED_ORIGINS = ["http://testserver", "http://localhost:8000"]
+
+EMAIL_BACKEND = "django.core.mail.backends.locmem.EmailBackend"
+
+PASSWORD_HASHERS = [
+    "django.contrib.auth.hashers.MD5PasswordHasher",
+]
+
+CACHES = {
+    "default": {
+        "BACKEND": "django.core.cache.backends.locmem.LocMemCache",
+        "LOCATION": "framehold-tests",
+    }
+}